Fix dolines loop-condition when inserting text
There was an error in dolines loop-condition, where positions were compared to a
static 'max position representing the end-position at the *start* of execution.
When characters were added to the buffer, the actual end-position was pushed
forward past 'max, so the loop would end early.

This is fixed by using a marker for the max position rather than a static
position.

Additionally, to fix border cases and ensure all lines are processed whether or
not the buffer ends with an empty line, forward-line is used for incrementing
and (point) is compared to the marker position.