Remove `my/evil-select-region-operator` (redundant)
There is already a function `evil-operator-range` that reads a motion and
returns a `(BEG END)` list (it can also take an optional argument to return
`(BEG END TYPE)`.

This can be used in place of `my/evil-select-region-operator`, and is
preferable since it doesn't require `call-interactively` to work.

There is also an off-by-one error in `my/evil-select-region-operator`:
the END value is one too high. The END value returned by `interactive` is
already one past the endpoint of the motion (as desired), but was then
erroneously incremented. This error is moot, now.

The operator `my/evil-replace-in-region` is converted to use
`evil-operator-range`.