my/dactyl-goto-command/function: trim spaces around input string
This is particularly useful for specifying an empty string (to goto the first
occurring command/function. Since these functions use a default for
'read-string, you can't enter an empty string when there is a valid symbol at
point. This change allows you to enter a space which will be converted to the
empty string.